Break tests to many modules and fix server test await problem
Earlier test_server reported not being awaited, since server.init() was not
mocked.
